refresh button working
