+++ src/Event/FindVenues.jsx
@@ -1,18 +1,10 @@
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import readAll from './../store/middleware/crud/readAll';
+import { useSelector } from 'react-redux';
 import VenueCard from './VenueCard';
 
 const FindVenues = () => {
-  const dispatch = useDispatch();
-
-  const user = useSelector(state => state.user.user);
   const venueList = useSelector(state => state.venue.userVenues);
 
-  useEffect(() => {
-    dispatch(readAll(user.token, 'venue'));
-    dispatch(readAll(user.token, `event/user/${user.id}`));
-  }, []);
+
 
   return (
     <>
